Project 3/Project 3/requirements.txt
tornado
mysqlclient
python-magic
bcrypt
pycrypto
pymongo
